Update NSS to 3.52.3 (Pre NSS-GYP)

+++ security/nss/lib/freebl/aeskeywrap.c
@@ -9,8 +9,6 @@
 #include "stubs.h"
 #endif
 
-#include <stddef.h>
-
 #include "prcpucfg.h"
 #if defined(IS_LITTLE_ENDIAN) || defined(SHA_NO_LONG_LONG)
 #define BIG_ENDIAN_WITH_64_BIT_REGISTERS 0
@@ -201,7 +199,7 @@ set_t(unsigned char *pt, unsigned long t)
 static void
 encode_PRUint32_BE(unsigned char *data, PRUint32 val)
 {
-    size_t i;
+    int i;
     for (i = 0; i < sizeof(PRUint32); i++) {
         data[i] = PORT_GET_BYTE_BE(val, i, sizeof(PRUint32));
     }
@@ -211,7 +209,7 @@ static PRUint32
 decode_PRUint32_BE(unsigned char *data)
 {
     PRUint32 val = 0;
-    size_t i;
+    int i;
 
     for (i = 0; i < sizeof(PRUint32); i++) {
         val = (val << PR_BITS_PER_BYTE) | data[i];
